pyproject.toml
[tool.poetry]
name = "spotibot"
version = "0.1.0"
description = ""
authors = ["Shahriar <me@shahriar.ee>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telebot = "^0.0.5"
savify = {git = "https://github.com/stschake/savify", rev = "feature/use-yt-dlp"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
